Break nrpyelliptic c library into modules
