+++ nrpy/infrastructures/BHaH/nrpyelliptic/__init__.py
@@ -1,0 +1,1 @@
+from . import auxiliary, diagnostics, initial_data, rhs_eval
