+++ nrpy/infrastructures/BHaH/nrpyelliptic/initial_data.py
@@ -1,0 +1,116 @@
+"""
+C function for setting the initial data to be used in hyperbolic relaxation.
+
+Authors: Thiago Assumpção; assumpcaothiago **at** gmail **dot** com
+         Zachariah B. Etienne; zachetie **at** gmail **dot* com
+"""
+
+from inspect import currentframe as cfr
+from types import FrameType as FT
+from typing import Union, cast
+
+import sympy as sp
+
+import nrpy.c_codegen as ccg
+import nrpy.c_function as cfc
+import nrpy.grid as gri
+import nrpy.helpers.parallel_codegen as pcg
+import nrpy.infrastructures.BHaH.simple_loop as lp
+
+
+# Define functions to set up initial guess
+def register_CFunction_initial_guess_single_point() -> Union[None, pcg.NRPyEnv_type]:
+    """
+    Register the C function for initial guess of solution at a single point.
+
+    :return: None if in registration phase, else the updated NRPy environment.
+    """
+    if pcg.pcg_registration_phase():
+        pcg.register_func_call(f"{__name__}.{cast(FT, cfr()).f_code.co_name}", locals())
+        return None
+
+    includes = ["BHaH_defines.h"]
+
+    desc = r"""Compute initial guess at a single point."""
+    cfunc_type = "void"
+    name = "initial_guess_single_point"
+    params = r"""const commondata_struct *restrict commondata, const params_struct *restrict params,
+    const REAL xx0, const REAL xx1, const REAL xx2,  REAL *restrict uu_ID, REAL *restrict vv_ID
+"""
+    body = ccg.c_codegen(
+        [sp.sympify(0), sp.sympify(0)],
+        ["*uu_ID", "*vv_ID"],
+        verbose=False,
+        include_braces=False,
+    )
+    cfc.register_CFunction(
+        includes=includes,
+        desc=desc,
+        cfunc_type=cfunc_type,
+        CoordSystem_for_wrapper_func="",
+        name=name,
+        params=params,
+        include_CodeParameters_h=True,
+        body=body,
+    )
+    return cast(pcg.NRPyEnv_type, pcg.NRPyEnv())
+
+
+def register_CFunction_initial_guess_all_points(
+    OMP_collapse: int,
+    enable_checkpointing: bool = False,
+) -> Union[None, pcg.NRPyEnv_type]:
+    """
+    Register the initial guess function for the hyperbolic relaxation equation.
+
+    :param enable_checkpointing: Attempt to read from a checkpoint file before generating initial guess.
+    :param OMP_collapse: Degree of OpenMP loop collapsing.
+
+    :return: None if in registration phase, else the updated NRPy environment.
+    """
+    if pcg.pcg_registration_phase():
+        pcg.register_func_call(f"{__name__}.{cast(FT, cfr()).f_code.co_name}", locals())
+        return None
+    includes = ["BHaH_defines.h", "BHaH_function_prototypes.h"]
+
+    desc = r"""Set initial guess to solutions of hyperbolic relaxation equation at all points."""
+    cfunc_type = "void"
+    name = "initial_data"
+    params = (
+        "commondata_struct *restrict commondata, griddata_struct *restrict griddata"
+    )
+    uu_gf_memaccess = gri.BHaHGridFunction.access_gf("uu")
+    vv_gf_memaccess = gri.BHaHGridFunction.access_gf("vv")
+    body = ""
+    if enable_checkpointing:
+        body += """// Attempt to read checkpoint file. If it doesn't exist, then continue. Otherwise return.
+if( read_checkpoint(commondata, griddata) ) return;
+"""
+    body += r"""for(int grid=0; grid<commondata->NUMGRIDS; grid++) {
+  // Unpack griddata struct:
+  params_struct *restrict params = &griddata[grid].params;
+#include "set_CodeParameters.h"
+  REAL *restrict xx[3];
+  for (int ww = 0; ww < 3; ww++)
+    xx[ww] = griddata[grid].xx[ww];
+  REAL *restrict in_gfs = griddata[grid].gridfuncs.y_n_gfs;
+"""
+    body += lp.simple_loop(
+        loop_body="initial_guess_single_point(commondata, params, xx0,xx1,xx2,"
+        f"&{uu_gf_memaccess},"
+        f"&{vv_gf_memaccess});",
+        read_xxs=True,
+        loop_region="all points",
+        OMP_collapse=OMP_collapse,
+    )
+    body += "}\n"
+    cfc.register_CFunction(
+        includes=includes,
+        desc=desc,
+        cfunc_type=cfunc_type,
+        name=name,
+        params=params,
+        include_CodeParameters_h=False,
+        body=body,
+    )
+    return cast(pcg.NRPyEnv_type, pcg.NRPyEnv())
